Hu 18

+++ src/main/java/com/pragma/powerup/plazamicroservice/domain/model/Category.java
@@ -1,21 +1,21 @@
 package com.pragma.powerup.plazamicroservice.domain.model;
 
 public class Category {
-    private int id;
+    private Long id;
     private String name;
     private String description;
 
-    public Category(int id, String name, String description) {
+    public Category(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
